Enable tox testing
Remove mock module to avoid collision with builtin mock module name. Moved Client class inside into agithub_test
