+++ tox.ini
@@ -1,0 +1,19 @@
+[tox]
+envlist = py27, py37, flake8
+
+[travis]
+python =
+    3.7: py37
+    2.7: py27
+
+[testenv:flake8]
+basepython = python
+deps = flake8
+commands = flake8 agithub setup.py
+
+[testenv]
+setenv =
+    PYTHONPATH = {toxinidir}
+deps = .[test]
+commands =
+    pytest {posargs}
